Move FederationStorageIndexKey to federation package
