+++ rskj-core/src/main/java/co/rsk/peg/federation/FederationStorageIndexKey.java
@@ -1,4 +1,4 @@
-package co.rsk.peg.storage;
+package co.rsk.peg.federation;
 
 import org.ethereum.vm.DataWord;
