+++ rskj-core/src/test/java/co/rsk/peg/federation/FederationStorageProviderImplTests.java
@@ -22,7 +22,7 @@ import java.util.stream.Stream;
 import static co.rsk.bitcoinj.core.NetworkParameters.ID_TESTNET;
 import static co.rsk.bitcoinj.core.NetworkParameters.ID_MAINNET;
 import static co.rsk.peg.federation.FederationFormatVersion.*;
-import static co.rsk.peg.storage.FederationStorageIndexKey.*;
+import static co.rsk.peg.federation.FederationStorageIndexKey.*;
 import static co.rsk.peg.BridgeSerializationUtils.serializeElection;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.core.RskAddress;
@@ -38,7 +38,6 @@ import co.rsk.peg.PegTestUtils;
 import co.rsk.peg.bitcoin.BitcoinTestUtils;
 import co.rsk.peg.constants.BridgeConstants;
 import co.rsk.peg.constants.BridgeMainNetConstants;
-import co.rsk.peg.storage.FederationStorageIndexKey;
 import co.rsk.peg.constants.BridgeRegTestConstants;
 import co.rsk.peg.federation.constants.FederationConstants;
 import co.rsk.peg.storage.BridgeStorageAccessorImpl;
